requirements.txt
alembic==1.4.2
altgraph==0.17
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
DateTime==4.3
Django==3.1
djangorestframework==3.12.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
install==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
netifaces==0.10.9
numpy==1.18.5
psutil==5.7.0
pydantic==1.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.16
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
